Most of functionality moved to BusInfo parent abstract class. Now child classes only have to implement getRawResult().
