+++ src/MPKList.java
@@ -1,0 +1,7 @@
+package PACKAGE_NAME;
+
+/**
+ * Created by umat on 02.05.17.
+ */
+public class MPKList {
+}
